Do something in the default DUContext::createNavigationWidget implementation.
When a valid declaration is given, we now return a navigation widget
with a declaration navigation context set. The names (Abstract...)
are very poorly chosen.

This makes it simpler for developers of new language plugins to
see something without first having to implement this method which
requires special DUContext overloads and such...

Generally, this method should probably be moved to the ILanguageSupport
interface I think...