Don't use WorstPriority in ParseProjectJob to make it possible to create jobs with worse
Php needs to reschedule documents for parsing if unknown symbols are encountered. This is
done by adding them to the background parser with a lower priority.

However this doesn't work on initial parsing when the project is opened as there all are
added with WorstPriority - and so can't get added with worse priority.

This patch makes that possible by using another lower priority in ParseProjectJob.

REVIEW: 108527