Simplify Code: build types directly in the DeclarationBuilder.
While the context builder is decoupled from the decl builder to
be reused in the use bulider, it does not make sense to also
decouple the typebuilder from the decl builder.

Merging both simplifies the code and makes it faster as well due
to less required locking, indirection and and lookups. We should
do that for all language plugins, really...