Simplify code and fix comment.
I think it should be fine for now to search starting from the local
context and let the default code bubble up to the parent contexts
as needed. At least I don't see the need for the complicated Ruby
code here.

If in future we hit cases where this is not enough we can still
introduce custom code but that should be accompanied by proper
unit tests then.