Do not hold a reference to the thread local data.
Apparently there are cases where the current thread changes between
construction of the memory pool and later calls to allocateBlock.
So far I could not figure out where this happens as the Parser e.g.
which showed up in the backtraces, was always constructed on the stack
inside the parse job's thread.

Anyhow, this patch here removes the references and simply asks for the
local storage whenever required. The benchmarks did not show any speed
impact here which is good. The crashes I had are also gone.