Avoid crashes when changing the selected builddir
The builddir-change triggers the clear() function which resets the paths
model. This in turn triggers the currentIndexChanged of the combobox with
a negative index. Since our path model should always have at least the
undeletable root-entry, we can safely try to use the first-entry when
this happens.

Keep the assert though to catch cases where somehow the code manages to
run while no root-entry is available.