Make it possible to kill DVcsJobs.
This is a safety measure, if a DVcsJob would be running and we close
KDevelop we cannot properly stop it. Then if result comes in we
might try to access our platform API in an invalid state.

This is the same reason why we see issues such as shown in bug
309715 - there for the CTestFindJob.