Make the title of the "Import / open Project" dialog more obvious
For some users, it was not at all clear that they could simply
open any folder as a project. I have answered the question "how to
convert a folder with source files to a KDevelop project" like
twenty times now, here's an example:
http://forum.kde.org/viewtopic.php?f=251&t=109911#p258755
With this change, if you're confused and read the title of the dialog
it should clear things up.