Increase the item repository version.
I think that I might have introduced an incompatiblity in the
recent changes to e.g. the SetDataRepository. This in turn then
resulted in strange crashes on startup like shown in bug 314031.

Lets keep that open for now but I think these crashes on startup in
KDevelop::ItemRepositoryRegistry::registerRepository should now be
resolved...

CCBUG: 314031