Display the current branch by the project name in the vcs changes view
Although I planned to implement it first for the project manager, I saw
that it involved quite some changes in the project model, so I decided to
put it here first and we can implement it in the project manager afterwards.

The patch adds in the IBranchingVersionControl interface a method to
require notifications when a branch has changed and a signal for those
notifications. Also it implements it in the git plugin by tracking changes
in the .git/HEAD file.