Remove crazy stuff from DVcsPlugin
The helper is already a QObject, it will be deleted with the parent
The empty method is not that much of a bad idea, but it's neither the
place nor the form (no i18n, no further information). The plugin can
implement something better in less than the time I spent writing this
commit message.