Improve the VcsEventsWidget
Move the logic of the log fetching from the widget to the model. Then from
the model we can do nice things like fetching the log by parts.
Then we fetch the log by parts, starting by 100 entries in an exponential
scale (100, 100, 200, 400, ...) so that it doesn't get too tough to get
very far in the past.
This will be useful because we won't need to make all the log processing
at once and we'll be able to use the widget before having it all.