Improve usability of the popup view
This change makes the project name more visible and provides a project-
relative path. In addition the size of the view is adjusted to give away
more of this valuable information to the user.

Initial idea provided by Jarosław Sierant <jaroslaw.sierant@gmail.com>

Edited by me to avoid the manual truncation and to make the view always
smaller than the mainwindows editor area since it looks somewhat strange
with a small mainwindow on a big screen.