Allow opening session by project on command line
If a session is not named, the first matching project file will be
opened instead. If any session has a name that is also used by a
project, that session will be opened. This makes it possible to open
unnamed sessions that are visible in -l without having to use the UUID.

REVIEW:109213