Simplify the code the the ProjectManagerView
The ProjectManagerView has a proxy model (and in a next commit it will have
two of them) and we have lots of code depending on it having a proxy model.
This patch cleans this up by adding two new roles in the ProjectModel: the
ProjectItemRole and the ProjectRole and a magic function
ProjectTreeView::mapFromSource() that will do the jumps regardless how many
proxies we have.

Also it improves the readability of the code (about 60 lines of code less)
and this also makes me happy.