Fix error if the first argument to the build command contains an '='.
splitArgs with AbortOnMeta will return a FoundMeta error if the first
word contains an equals sign, because it thinks it's a shell variable
assignment, as in "LANG=C make" or similar. We were passing the list
of arguments of the build command to KShell::splitArgs, without
including the command itself. Thus, if the first argument (*after*
the command name!) had an equals sign, like in command="make"
arguments="V=1", KDevelop was saying "the given arguments would need
a real shell", which isn't true.

This patch fixes the problem by passing the command name to splitArgs
too (and then removing the command name from the returned split list).

REVIEW: 106236