Refactor file/project templates lookup code.
We used to register a custom resource type and queried that for
the template preview, description and archive files.

The custom resource type could potentially trigger a qFatal in
KStandardDirs (which I'll fix in a patch to KDELibs separately).
But, according to DFaure, one should not use custom resource types
anyways - esp. since that feature is gone in KF5 anyways.

So now we instead always use the 'data' type with the different
folders as filters. Works nicely but I think we should merge more
of this lookup code into a central class which should allow us to
share quite a bit of code. Something for 4.6?