Fix issue in grepview plugin
It was impossible to replace "\n" with "\n\n" before.

I really don't know what the code path replacing '\\' by '\\\\'
was supposed to do. It replaced characters in the replacement string in
case the input regexp was marked with QRegExp::Wildcard. But I have no
clue why the replacement string should depend on that.

Add tests.

BUG: 301362
REVIEW: 109554