Fix crash when adding a new build config
I thought I had this fixed already some time ago, but apparently not. When
adding a new build config the resetting of the paths emits a model-reset
and that triggers the currentIndex-changed of the path combobox. However
the new index is -1 at that point since the data is empty. Since the callers
all make sure to re-add the "root" entry after clear, lets simply not update
our internal data from inside clear to avoid the index-changed being
emitted.