Do not call QApplication::processEvents in KDevSplashScreen::progress.
This is hazardous and confuses the eventloop tracking of KJobs
apparently, as can be seen by KDevelop not shutting down properly
since some KJob is still KGlobal::ref'ed and thus prevents the final
shutdown of the QApplication.

I found this call in the backtraces of the creation of the "leaked"
KJobs thus assumed it is the reason. And indeed, removing it fixes
the shutdown problem for me. Please confirm!