Rename some .desktop.cmake back to .desktop.
d043dcdb changed all .desktop files to be processed by CMake,
so that X-KDevelop-Version is set automatically. But some files
don't have that key at all, and some are used by unit tests
which weren't modified to look for the changed file name.
So here I revert the change for .desktop files where it wasn't necessary.