Rename KDevMimeTypeItem to KDevCategoryItem
It doesn't show mimetypes anymore, rename it with a bit of slack for
other categorization ideas.

Also allow us to easily change the top-level branch indicators for
directories from the expand indicator to directory icons. I personally
prefer the more minimal approach, but maybe people like icons more?
Note: Using icons would allow us to ditch a bunch of lines of custom
painting code in the delegates, which is tempting. :>