Don't rely on output model to read stdout/err.
The outputmodel is not instant, thus unreliable when looking for the
contents of the stdout/err buffer after a process has finished.

Instead use store an (implicitly shared copy) of the plain data as
received from the process line maker.

This fixes insertion of process output or errors into documents using
the external scripts plugin.