Enable having different IExecutePlugin instances supported from gdb plugin
Right now we have an abstraction to have a launch configuration that can
be executed that the GDB plugin attaches to it to debug it. What this patch
does is to look not for one but for all plugins implementing IExecutePlugin
and initializing them to be used.
This will be used by the ExecutePlasmoid plugin, to make it possible to have
them debugged using gdb.