Fix initialization of the project controller in ctestfindsuitestest (and probably others)
When the unit test was started, the projects were restored from the session
and it was opened as well. It resulted in some code called by the event loop
opening the project as well. This asserted because the project was already
being opened.
This patch puts the project url as currentlyOpening before, so that when
the second call is triggered, it finds out that the project is being opened
already, so we don't hit the assert.