Improve how we react to missing core files
Right now, I had this test not passing because systemd puts the core files
in some weird place and I even have to be root to fetch those. It's clearly
something that this unit test is not managing, so I'd say it's fine to skip
it on systems like mine until somebody finds the time to figure it out.
I added some information so that the next person to look into this knows
what's going on.