Set a minimum size for the documentswitcher view
With only two files the popup is so small that only one of them is visible
which is very irritating. Even with 3 or 4 files it looks very smallish, so
make it at least as big as 6 entries would require it. This adds a bit of
whitespace, but looks much less hacky.