Preserve relative paths in unresolved include assistant
The unresolved include assistant include relative paths which are
resolved from the storage directory. However, when reopening the
assistant, all relative paths would be turned into absolute ones. Fix
this by converting relative paths in a dedicated method of
CustomIncludePathsSettings and use this one in places where absolute
paths are desirable, leaving the assistant with the original relative
paths for editing.

REVIEW: 110295