Revert "Remove the .sh extension from scripts installed in prefix/bin"
This reverts commit c883b79e2a3654866f0c6ae4b75fe96402136533.

Unbreaks installing kdevelop as normal user and probably unbreaks
using the script as well as KDE4_LIBEXEC_INSTALL_DIR is not in users PATH
on most systems.

Lets keep bin/ for now, if packagers absolutely require the scripts in
libexec then the code calling them needs to be adjusted as well as making
sure libexec is selected inside the cmake install prefix and not /usr or
so (whatever is kdelibs install prefix).

BUG: 319915