Optimize: Cache CMakeParserUtils::initialVariables result.
This is only dependent on the globally accessible cmake binary found
via KStandardDirs::findExe and thus returns the same whenever you
call it.

Reason why we want to cache it is that it takes about 1s on my machine
and is queried at least once for every project. Considering that I
usually run a session with 5-10 projects opened this accounts for
5-10s of time wasted querying the same data from cmake over and over
again. Not anymore with this patch - yay!

REVIEW: 110580