Remove the need to install the google sparsehash headers
They were only needed because of the KDevelop::DUContextDynamicData class
being installed through the ducontextdata.h file which had both this class
and DUContextData. Now we have a separate file for DUContextDynamicData and
we are not installing it because it doesn't seem to be needed outside
kdevplatform.

CCMAIL: kdevelop-devel@kde.org