Transform ranges properly using the revision they actually belong to.
This is imo basically a hack to circumvent the bad RangeInRevision API
without rewriting everything...

The RangeInRevision is actually just blindly assuming that it belongs
to the revision the file was parsed in the last time. But if the file
is reparsed in between (due to changes) the ranges might be wrong. Due
to that we must store the ranges and their actual revision... Stupid!

Imo an API called RangeInRevision should _know_ which revision it
belongs to... Meh!

BUG: 295707