Refactored the QtHelp loading so that it is not synchronous
As it was being done at the moment, we were calling synchronously the qmake
process from the plugin constructor. This was bad and had problems, like
the one discussed in the bug.
Some logic for the lookup is being removed as well, but it seemed like an
overkill to me, the first qmake should be already working. If that's not
the case we can iterate over it.

BUG: 310703