Install the Grantlee filter into the right location.
This is similar to what was done to Skrooge in commit 14baa01.

Grantlee plugins, such as kdev_filters, are Qt plugins, and thus it makes
sense to install it into ${PLUGIN_INSTALL_DIR} instead of installing into
the same directory Grantlee uses for its own filters.

It's also worth pointing out that people are expected to have the
QT_PLUGIN_PATH environment variable correctly set for the filter to be
found by Grantlee.

REVIEW:	111155