Hide empty folders after filtering
Now when using the filter bar in the projects panel, the folders that
contain no items matching the filter are hidden, so you don't have to
manually search through all the folders to see if there are items or
not. I know this is slower than the previous code, however I tried this
on a big project (Qt) and the  impact on performance was unnoticeable.

Additionally, Aleix suggested to change the label next to the items when
the filter is enabled, so I did it too in this patch. Now it displays
the number of hidden items, instead of the visible ones.

REVIEW:111537