Fix macro expansion when the arguments are split over multiple lines.
When a macro is used that requires arguments and those arguments are
split over multiple lines kdevelop is not able to expand the macro.
The expander respects only the current line and therefore the parser
does not find the corresponding close brace. In this fix the next
line is added to the argument list until the close brace is found or
EOF is reached.

REVIEW: 111007