First step towards a generic, shared project filter KCM.
This adds a common IProjectFilter interface and a generic plugin
which implements it. The implementation is just a copy of the old
code that was written for the generic manager.

TODO: actually query the filters and use that in the project managers.
TODO: cleanup the code
TODO: optimize performance, esp. cache the QRegExp