Follow renames when showing git history.
By default git log (used to get history info in KDevelop) does not
follow the file moves or renames. The patch adds the --follow
argument to the git command to get the complete history.

I got bitten by this, because it was not possible to identify all
committers to a file by using KDevelop.

REVIEW: 109952