Make the project filter patterns immutable.
This also requires a change to when IProjectController emits
the projectAboutToBeOpened signal, which was up to now completely
useless (raw ptr without any initialization done, no way to find
name, url, config values, ...).

The change itself will be helpful in making it threadsafe in a more
efficient way I hope.

The problem will be user-induced changes to the filter while the
model is being reloaded (e.g. due to external changes).