Extend ProjectFilterManager API to enable threadsafe filtering.
The API itself is still not threadsafe, it just now offers more
functionality based on which you can implement a properly threadsafe
project filtering. This will be done in the CMakeManager with an
upcoming commit.

The way you should handle this is by getting a list of threadsafe
project filters (via filtersForProject()). Then use this list directly
in your background thread, which is then a threadsafe operation.