Implement project filtering for the cmake manager in a threadsafe way.
Operate directly on a list of thread safe project filters in the
background thread. ::parse() needs a hack to ensure that this list
is queried in a threadsafe manner, as there is currently no proper
place in the API to do this.

In the future, a custom import job should be created which then reads
the list of filters before going of into the background thread.

I've tested this with a session containing all KDEPIM projects. The
slowdown compared to before is negleglible and I did not encounter any
crashes.