Add a reviewboardrc for easier review-creation
As far as I understood http://techbase.kde.org/Development/Review_Board
its still a bit cumbersome to create a review request due to needing the
git:// urls etc. However with this minimal rc-file I was able to just
tell post-review to guess summary+description and it figured out
everything else. So lets add this to the repository.