Optimize: Greatly speed up painting of the QuickOpen list of files.
The bottleneck is the QPixmap creation from a KIcon. We now cache
the KIconLoader::Small icon. In my tests this is a difference of
day and night performance wise. Where scrolling in a longish list
easily took 100% CPU and felt sluggish before, it now is very smooth.

This allows us to remove the scrollbar hack, where we disabled it
when too many items where displayed. It's not neccessary anymore.