Shutup Clang warnings about alignment increases.
The whole structure of our on-disk serialization does not take
alignment into account very well. This should be improved in a
proper rewrite, but for now there is not much we can do. Also note
that since we dump data from memory directly on disk, the data there
is aligned as if in memory. The issue is with the additional meta-data
on top, which might introduce misalignments.

Still, most cases where the warning is emitted its probably not an
issue. If we instead use reinterpret_cast there, this actually removes
the warning. In the future we should then double-check for alignment
issues in the deserialization wrapper.