Add a test for missing support of base class init'ing via typedefs.
I tried to implement it but it's far from easy as far as I can see.
The NameVisitor which is called from ExpressionVisitor::visitName
has the "DontSearchInParents" flag set when m_memberAccess is true.

This then fails to find the typedef declaration of the base class
which is required to create the proper use there...

Anyone an idea how to implement that properly?