Simplify the filter configuration, remove "MatchOn".
This information can easily be deduced from the pattern and thus
gives a perfect opportunity for a simplified configuration UI.
It should thus also be closer to what a user expects and knows from
e.g. .gitignore and svnignore files.

Thanks to Andreas Pakulat for the suggestion.

The Filter struct now should be seen as a 'compiled' user filter,
whereas SerializedFilter holds the raw data the user entered and
thus is what is used in the configu UI.