Add action to exclude project items to the ProjectItemContext menu.
This then adds filters with the explicit relative path to the project
configuration and notifies the consumers that the filters for the
given project have changed. A message box with a "dont show again"
action is shown afterwards, explaining how to user can undo this
action by modifying the project filter settings.