Add support for drag'n'drop moving of filters.
Note to self: More work than I assumed and learned a lot of what
is required to get this done using Qt's model/view pattern. I find
it especially odd that setData is not called on all columns. By
overwriting itemData/setItemData we can make it work though. Also
odd is the drop behavior at the end of the view which tries to insert
at an invalid index position... Also, if one returns false for
e.g. insertRow the removeRow and/or setData stuff is still being
called - meh!