Another try at me beautifying the new area display.
Use a toolbutton again - yay! This time in it's "flat" design though.
I really prefer it for it's reduced margin on the left and the better
placement of the arrow menu indicator at the bottom left. The push
button has the arrow directly next to the next which looks really
cramped. Since it's still a "flat" button, nothing else changes.