Do session name squeezing in the menu action generator, not the session
This way, we get small names for the menus, but still the full names for
the session picker. This is relevant because it again allows to search
for a project name in the session picker and find all sessions with that
project. Until now, it might have been cut off.