Fix default when reading the number of jobs
Using different defaults for the .kcfg file and the code reading the
value leads to a problem when the user configures the project to the
'default' value (2 in this case). Saving the default-value will simply
remove it from the config file, which leaves the reading-code to use its
own default.

In this case this means, a user configuring to use 2 jobs for make, will
end up with just 1 as the reading-code uses that as default. When the
user selects 3 or more for jobs, it suddenly magically starts working.

For now a quick-fix to use '2', a more permanent solution would be to
use the generated file from the makejob as well. That could be a bit
more involved though since the setup of the class is a bit complicated
(see projectkcmodule.h).