Add convenience functions taking Path for Project helpers.
These helpers are a glue to other libraries which do not use
the Path api (yet?). Thus the URL-taking versions are fine but we
want to allow users of our KDevplatform Project API to always use
Path and minimize the usage of .toUrl() there.

Note that the conversions here are OK as its not done often and
the cost is dwarfed by actually doing KIO operations and showing
dialogs etc. pp.