Remove unused Run > profile menu entry
Currently we had the menu entry but nobody using it. Technically it's
correct because whenever we'd get a profiler it would work automatically
but I'd say that also having the option but not being able to use
it is nothing but frustrating.

Commenting out the code until somebody puts some care into profiling
integration.

BUG: 272534