Remove QPropertyDeclaration.
This was never supposed to be a declaration, you cannot interact
with that from C++ code after all. In the future, we might want
an additional "something" for proper integration with e.g. QML.

Right now, this code is causing more problems than is worth it,
like issues in renaming or context browsing, since often the
property name clashes with the getter function. Only the latter
is actually useful in a C++ context.

Instead, we now report proper uses for the getter/setter/notify/...
functions in the Q_PROPERTY line.