Increase static item repository version.
This is required, as you can otherwise run into assertions after
updating KDevelop. The removal of the QPropertyDeclaration there
invalidates the cache and one cannot load the stored declarations.
The code is apparently too frail to cope with that and thus asserts.

To prevent any issues here, lets just increase the repo version.