C++11: Add implicit using directive for inline namespaces.
This fixes association of uses etc. to contents in inline namespaces,
which is the case for e.g. most of the stuff in std::chrono.

Note: Code completion of items in an inline namespace nested in
another namespace (such as is the case for stuff in std::chrono::_V2)
does not show up properly yet.

BUG: 321104