Paint full splash screen in QML.
This is especially nice since it reduces the maintenance whenever
we increase the version number. Now we simply show the version
of the compiled KDevelop application which is much better than having
to update a pixmap everytime.

And if the QML fails to load for some reason, well its a bug and
should be fixed. Having a pixmap fallback for that does not make
sense in my eyes.