Remove proxy-model filtering from ProjectManagerView/ProjectProxyModel
This is now handled in a much more efficient way via the ProjectFilter
plugin. The dynamic filtering in the project view is thus obsoleted.

Thus: If you want to exclude stuff from your project, use the filter
plugin. If you want to quickly open a file for a given pattern, use
quickopen.