Don't show a message box when a plugin couldn't be loaded
IPlugin has a hasError method that is used to check if the plugin could
initialize itself. It's used to make sure that the runtime dependencies
are met, such as checking that git, cmake and ninja are installed.

Having a message box for those is too intrusive, given that if the user
needs the tool and doesn't have them installed, he'll have a bigger
problem than not having the plugin in KDevelop. Instead, a kWarning() is
provided, so we can tell our users to check the console output if one of
these problems happen.

CCMAIL: kdevelop-devel@kde.org