Cleanup code and fix building of single files in CMake/Make projects.
Temporary model items which are never added to the model are not
supported by the MakeJob anymore, as it only stores a QPersistentModel
Index internally, instead of the raw pointer.

We fix this by using the MakeBuilder API instead which works and is
much cleaner as well imo. Sadly, that API is not part of the Ninja
Builder though, thus it won't work there for now.